change conversion time to uint8_t

+++ esphome/components/bmp581/bmp581.h
@@ -83,7 +83,7 @@ class BMP581Component : public PollingComponent, public i2c::I2CDevice {
   }
   void set_pressure_iir_filter_config(IIRFilter iir_pressure_level) { this->iir_pressure_level_ = iir_pressure_level; }
 
-  void set_conversion_time(uint16_t conversion_time) { this->conversion_time_ = conversion_time; }
+  void set_conversion_time(uint8_t conversion_time) { this->conversion_time_ = conversion_time; }
 
  protected:
   sensor::Sensor *temperature_sensor_{nullptr};
@@ -97,7 +97,7 @@ class BMP581Component : public PollingComponent, public i2c::I2CDevice {
 
   // Stores the sensors conversion time needed for a measurement based on oversampling settings and datasheet (page 12)
   // Computed in Python during codegen
-  uint16_t conversion_time_;
+  uint8_t conversion_time_;
 
   // Checks if the BMP581 has measurement data ready by checking the sensor's interrupts
   bool check_data_readiness_();
